% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{Combinations-class}
\alias{Combinations-class}
\title{Combinations class}
\arguments{
\item{n}{integer: number of total items;
\code{n} may be implicitly determined by \code{x} and \code{f} if missing}

\item{r}{integer: number of selected items}

\item{x}{a vector: optional labeled vector}

\item{f}{an integer vector: frequency for each item}

\item{replace}{with/without replacement}
}
\description{
\preformatted{
Combinations$new(n, r, x = NULL, f = NULL, replace = FALSE)
}
}
