% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{Combinations}
\alias{Combinations}
\alias{icombinations}
\title{Combinations iterator}
\usage{
icombinations(
  x = NULL,
  k = NULL,
  n = NULL,
  v = NULL,
  freq = NULL,
  replace = FALSE,
  skip = NULL
)
}
\arguments{
\item{x}{an integer or a vector, will be treated as \code{n} if integer; otherwise, will be treated as \code{v}.
Should not be specified together with \code{n} and \code{v}.}

\item{k}{an integer, the number of items drawn, defaults to \code{n} if \code{freq} is \code{NULL} else \code{sum(freq)}}

\item{n}{an integer, the total number of items, its value may be implicitly deduced from \code{length(v)} or \code{length(freq)}}

\item{v}{a vector to be drawn, defaults to \code{1:n}.}

\item{freq}{an integer vector of item repeat frequencies}

\item{replace}{an logical to draw items with replacement}

\item{skip}{the number of combinations skipped}
}
\description{
This function returns a \link{Combinations} iterator for iterating
combinations of \code{k} items from \code{n} items. The iterator allows users to fetch the next
combination(s) via the \code{getnext()} method.
}
\details{
The \code{Combinations} class can be initialized by using the convenient wrapper \code{icombinations} or
\preformatted{
Combinations$new(n, k, v = NULL, freq = NULL, replace = FALSE)
}

\preformatted{
getnext(d = 1L, layout = NULL, drop = NULL)
collect(layout = "row")
reset()
}
\describe{
\item{d}{number of fetched arrangements}

\item{layout}{if "row", "column" or "list" is specified, the returned value would be a
"row-major" matrix, a "column-major" matrix or a list respectively}

\item{drop}{vectorize a matrix or unlist a list}
}
}
\examples{
icomb <- icombinations(5, 2)
icomb$getnext()
icomb$getnext(2)
icomb$getnext(layout = "column", drop = FALSE)
# collect remaining combinations
icomb$collect()

library(foreach)
foreach(x = icombinations(5, 2), .combine=c) \%do\% {
  sum(x)
}
}
\seealso{
\link{combinations} for generating all combinations and \link{ncombinations} to calculate number of combinations
}
\section{Super classes}{
\code{arrangements::abstractiter} -> \code{arrangements::iter} -> \code{arrangements::Arrangements} -> \code{Combinations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Combinations-new}{\code{Combinations$new()}}
\item \href{#method-Combinations-reset}{\code{Combinations$reset()}}
\item \href{#method-Combinations-collect}{\code{Combinations$collect()}}
\item \href{#method-Combinations-getnext}{\code{Combinations$getnext()}}
\item \href{#method-Combinations-print}{\code{Combinations$print()}}
\item \href{#method-Combinations-clone}{\code{Combinations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="arrangements" data-topic="Arrangements" data-id="nextElem"><a href='../../arrangements/html/Arrangements.html#method-Arrangements-nextElem'><code>arrangements::Arrangements$nextElem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-new"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$new(n, k, v = NULL, freq = NULL, replace = FALSE, skip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-collect"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$collect(layout = "row")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-getnext"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-getnext}{}}}
\subsection{Method \code{getnext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$getnext(d = 1L, layout = NULL, drop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-print"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Combinations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Combinations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Combinations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
