% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{Compositions}
\alias{Compositions}
\alias{icompositions}
\title{Compositions iterator}
\usage{
icompositions(n, k = NULL, descending = FALSE, skip = NULL)
}
\arguments{
\item{n}{an non-negative integer to be partitioned}

\item{k}{number of parts}

\item{descending}{an logical to use reversed lexicographical order}

\item{skip}{the number of compositions skipped}
}
\description{
This function returns a \link{Compositions} iterator for iterating
compositions of an non-negative integer \code{n} into \code{k} parts or parts of any sizes.
The iterator allows users to fetch the next partition(s) via the \code{getnext()} method.
}
\details{
The \code{Compositions} class can be initialized by using the convenient wrapper \code{icompositions} or
\preformatted{
Compositions$new(n, k = NULL, descending = FALSE)
}

\preformatted{
getnext(d = 1L, layout = NULL, drop = NULL)
collect(layout = "row")
reset()
}
\describe{
\item{d}{number of fetched arrangements}

\item{layout}{if "row", "column" or "list" is specified, the returned value would be a
"row-major" matrix, a "column-major" matrix or a list respectively}

\item{drop}{vectorize a matrix or unlist a list}
}
}
\examples{
ipart <- icompositions(4)
ipart$getnext()
ipart$getnext(2)
ipart$getnext(layout = "column", drop = FALSE)
# collect remaining compositions
ipart$collect()

library(foreach)
foreach(x = icompositions(6, 2), .combine=c) \%do\% {
  prod(x)
}
}
\seealso{
\link{compositions} for generating all compositions and \link{ncompositions} to calculate number of compositions
}
\section{Super classes}{
\code{arrangements::abstractiter} -> \code{arrangements::iter} -> \code{arrangements::Arrangements} -> \code{Compositions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compositions-new}{\code{Compositions$new()}}
\item \href{#method-Compositions-reset}{\code{Compositions$reset()}}
\item \href{#method-Compositions-collect}{\code{Compositions$collect()}}
\item \href{#method-Compositions-getnext}{\code{Compositions$getnext()}}
\item \href{#method-Compositions-print}{\code{Compositions$print()}}
\item \href{#method-Compositions-clone}{\code{Compositions$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="arrangements" data-topic="Arrangements" data-id="nextElem"><a href='../../arrangements/html/Arrangements.html#method-Arrangements-nextElem'><code>arrangements::Arrangements$nextElem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-new"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$new(n, k = NULL, descending = FALSE, skip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-collect"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$collect(layout = "row")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-getnext"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-getnext}{}}}
\subsection{Method \code{getnext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$getnext(d = 1L, layout = NULL, drop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-print"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compositions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compositions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compositions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
