% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{Partitions}
\alias{Partitions}
\alias{ipartitions}
\title{Partitions iterator}
\usage{
ipartitions(n, k = NULL, distinct = FALSE, descending = FALSE, skip = NULL)
}
\arguments{
\item{n}{an non-negative integer to be partitioned}

\item{k}{number of parts}

\item{distinct}{an logical to restrict distinct values}

\item{descending}{an logical to use reversed lexicographical order}

\item{skip}{the number of partitions skipped}
}
\description{
This function returns a \link{Partitions} iterator for iterating
partitions of an non-negative integer \code{n} into \code{k} parts or parts of any sizes.
The iterator allows users to fetch the next partition(s) via the \code{getnext()} method.
}
\details{
The \code{Partitions} class can be initialized by using the convenient wrapper \code{ipartitions} or
\preformatted{
Partitions$new(n, k = NULL, descending = FALSE)
}

\preformatted{
getnext(d = 1L, layout = NULL, drop = NULL)
collect(layout = "row")
reset()
}
\describe{
\item{d}{number of fetched arrangements}

\item{layout}{if "row", "column" or "list" is specified, the returned value would be a
"row-major" matrix, a "column-major" matrix or a list respectively}

\item{drop}{vectorize a matrix or unlist a list}
}
}
\examples{
ipart <- ipartitions(10)
ipart$getnext()
ipart$getnext(2)
ipart$getnext(layout = "column", drop = FALSE)
# collect remaining partitions
ipart$collect()

library(foreach)
foreach(x = ipartitions(6, 2), .combine=c) \%do\% {
  prod(x)
}
}
\seealso{
\link{partitions} for generating all partitions and \link{npartitions} to calculate number of partitions
}
\section{Super classes}{
\code{arrangements::abstractiter} -> \code{arrangements::iter} -> \code{arrangements::Arrangements} -> \code{Partitions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Partitions-new}{\code{Partitions$new()}}
\item \href{#method-Partitions-reset}{\code{Partitions$reset()}}
\item \href{#method-Partitions-collect}{\code{Partitions$collect()}}
\item \href{#method-Partitions-getnext}{\code{Partitions$getnext()}}
\item \href{#method-Partitions-print}{\code{Partitions$print()}}
\item \href{#method-Partitions-clone}{\code{Partitions$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="arrangements" data-topic="Arrangements" data-id="nextElem"><a href='../../arrangements/html/Arrangements.html#method-Arrangements-nextElem'><code>arrangements::Arrangements$nextElem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-new"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$new(n, k = NULL, distinct = FALSE, descending = FALSE, skip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-collect"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$collect(layout = "row")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-getnext"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-getnext}{}}}
\subsection{Method \code{getnext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$getnext(d = 1L, layout = NULL, drop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-print"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Partitions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Partitions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Partitions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
