% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{Permutations}
\alias{Permutations}
\alias{ipermutations}
\title{Permutations iterator}
\usage{
ipermutations(
  x = NULL,
  k = NULL,
  n = NULL,
  v = NULL,
  freq = NULL,
  replace = FALSE,
  skip = NULL
)
}
\arguments{
\item{x}{an integer or a vector, will be treated as \code{n} if integer; otherwise, will be treated as \code{v}.
Should not be specified together with \code{n} and \code{v}.}

\item{k}{an integer, the number of items drawn, defaults to \code{n} if \code{freq} is \code{NULL} else \code{sum(freq)}}

\item{n}{an integer, the total number of items, its value may be implicitly deduced from \code{length(v)} or \code{length(freq)}}

\item{v}{a vector to be drawn, defaults to \code{1:n}.}

\item{freq}{an integer vector of item repeat frequencies}

\item{replace}{an logical to draw items with replacement}

\item{skip}{the number of combinations skipped}
}
\description{
This function returns a \link{Permutations} iterator for iterating
permutations of \code{k} items from \code{n} items. The iterator allows users to fetch the next
permutation(s) via the \code{getnext()} method.
}
\details{
The \code{Permutations} class can be initialized by using the convenient wrapper \code{ipermutations} or
\preformatted{
Permutations$new(n, k, v = NULL, freq = NULL, replace = FALSE)
}

\preformatted{
getnext(d = 1L, layout = NULL, drop = NULL)
collect(layout = "row")
reset()
}
\describe{
\item{d}{number of fetched arrangements}

\item{layout}{if "row", "column" or "list" is specified, the returned value would be a
"row-major" matrix, a "column-major" matrix or a list respectively}

\item{drop}{vectorize a matrix or unlist a list}
}
}
\examples{
iperm <- ipermutations(5, 2)
iperm$getnext()
iperm$getnext(2)
iperm$getnext(layout = "column", drop = FALSE)
# collect remaining permutations
iperm$collect()

library(foreach)
foreach(x = ipermutations(5, 2), .combine=c) \%do\% {
  sum(x)
}
}
\seealso{
\link{permutations} for generating all permutations and \link{npermutations} to calculate number of permutations
}
\section{Super classes}{
\code{arrangements::abstractiter} -> \code{arrangements::iter} -> \code{arrangements::Arrangements} -> \code{Permutations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Permutations-new}{\code{Permutations$new()}}
\item \href{#method-Permutations-reset}{\code{Permutations$reset()}}
\item \href{#method-Permutations-collect}{\code{Permutations$collect()}}
\item \href{#method-Permutations-getnext}{\code{Permutations$getnext()}}
\item \href{#method-Permutations-print}{\code{Permutations$print()}}
\item \href{#method-Permutations-clone}{\code{Permutations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="arrangements" data-topic="Arrangements" data-id="nextElem"><a href='../../arrangements/html/Arrangements.html#method-Arrangements-nextElem'><code>arrangements::Arrangements$nextElem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-new"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$new(
  n,
  k = NULL,
  v = NULL,
  freq = NULL,
  replace = FALSE,
  skip = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-collect"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$collect(layout = "row")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-getnext"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-getnext}{}}}
\subsection{Method \code{getnext()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$getnext(d = 1L, layout = NULL, drop = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-print"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Permutations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Permutations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permutations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
