% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{ncombinations}
\alias{ncombinations}
\title{Number of combinations}
\usage{
ncombinations(n, k, x = NULL, f = NULL, replace = FALSE, bigz = FALSE)
}
\arguments{
\item{n}{an integer, would be determined implicitly from `x` or `f` if missing}

\item{k}{an integer}

\item{x}{an optional vector indicating item labels}

\item{f}{an integer vector of item repeat frequencies}

\item{replace}{an logical to select with replacement}

\item{bigz}{an logical to indicate using \link[gmp:bigz]{gmp::bigz}}
}
\description{
Number of combinations
}
\seealso{
\link{combinations} for generating combinations and \link{icombinations} for iterating combinations
}
