% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{npermutations}
\alias{npermutations}
\title{Number of permutations}
\usage{
npermutations(n, k = n, x = NULL, f = NULL, replace = FALSE,
  bigz = FALSE)
}
\arguments{
\item{n}{an integer, would be determined implicitly from `x` or `f` if missing}

\item{k}{an integer}

\item{x}{an optional vector indicating item labels}

\item{f}{an integer vector of item repeat frequencies}

\item{replace}{an logical to select with replacement}

\item{bigz}{an logical to indicate using \link[gmp:bigz]{gmp::bigz}}
}
\description{
Number of permutations
}
