% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\docType{data}
\name{Partitions-class}
\alias{Partitions-class}
\alias{Partitions}
\title{Partitions class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Partitions
}
\description{
An R6 class of partitions iterator. \link{ipartitions} is a convenient wrapper for initializing the class.
}
\section{Signatures}{

\preformatted{
Partitions$new(n, k, x = NULL, f = NULL, replace = FALSE)
.$getnext(d = 1L, type = "r", drop = d == 1L)
.$collect(type = "r")
}
}

\seealso{
\link{ipartitions}
}
\keyword{datasets}
