% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{Permutations-class}
\alias{Permutations-class}
\title{Permutations class}
\description{
An R6 class of permutation iterator. \link{ipermutations} is a convenient wrapper for initializing the class.
}
\section{Usage}{

\preformatted{
Permutations$new(n, k, x = NULL, f = NULL, replace = FALSE)
...$getnext(d = 1L, type = "r", drop = d == 1L)
...$collect(type = "r")
}
}

\seealso{
\link{ipermutations}
}
