% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Permutations generator}
\usage{
permutations(n, k = n, x = NULL, f = NULL, replace = FALSE,
  type = "r")
}
\arguments{
\item{n}{an integer, would be determined implicitly from `x` or `f` if missing}

\item{k}{an integer}

\item{x}{an optional vector indicating item labels}

\item{f}{an integer vector of item repeat frequencies}

\item{replace}{an logical to select with replacement}

\item{type}{if "r", "c" or "l" is specified, the return results would be a
"row-major" matrix, "column-major" matrix or a list respectively}
}
\value{
a matrix if \code{type} is "r" or "c", a list if \code{type} is "l".
}
\description{
This function generates all the permutations of selecting \code{k} items from \code{n} items.
The results are in lexicographical order.
}
