% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangements.R
\docType{package}
\name{arrangements-package}
\alias{arrangements}
\alias{arrangements-package}
\title{arrangements: Fast Generators and Iterators for Permutations, Combinations, Integer
    Partitions and Compositions}
\description{
Fast generators and iterators for permutations, combinations,
    integer partitions and compositions. The arrangements are in
    lexicographical order and generated iteratively in a memory efficient manner. 
    It has been demonstrated that 'arrangements' outperforms most existing
    packages of similar kind. Benchmarks could be found at
    <https://randy3k.github.io/arrangements/articles/benchmark.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://randy3k.github.io/arrangements}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
