% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{Combinations-class}
\alias{Combinations-class}
\alias{Combinations}
\title{Combinations class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Combinations
}
\description{
An R6 class of combinations iterator. \link{icombinations} is a convenient wrapper for initializing the class.
}
\section{Initialization}{

\preformatted{
Combinations$new(n, k, x = NULL, f = NULL, replace = FALSE)
}
}

\section{Methods}{

\preformatted{
getnext(d = 1L, type = NULL, drop = d == 1L && is.null(type))
collect(type = "r")
reset()
}
\describe{
\item{\code{d}}{number of fetched arrangements}
\item{\code{type}}{if "r", "c" or "l" is specified, the returned value would be a
 "row-major" matrix, a "column-major" matrix or a list respectively}
\item{\code{drop}}{vectorize a matrix or unlist a list}
}
}

\seealso{
\link{icombinations}
}
\keyword{datasets}
