% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{Combinations-class}
\alias{Combinations-class}
\alias{Combinations}
\title{Combinations class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Combinations
}
\description{
An R6 class of combinations iterator. \link{icombinations} is a convenient wrapper for initializing the class.
}
\section{Methods}{

\preformatted{
Combinations$new(n, k, x = NULL, f = NULL, replace = FALSE)
.$getnext(d = 1L, type = "r", drop = d == 1L)
.$collect(type = "r")
}
}

\seealso{
\link{icombinations}
}
\keyword{datasets}
