% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{Combinations-class}
\alias{Combinations-class}
\title{Combinations class}
\description{
An R6 class of combinations iterator. \link{icombinations} is a convenient wrapper for initializing the class.
}
\section{Usage}{

\preformatted{
Combinations$new(n, k, x = NULL, f = NULL, replace = FALSE)
...$getnext(d = 1L, type = "r", drop = d == 1L)
...$collect(type = "r")
}
}

\seealso{
\link{icombinations}
}
