% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{icombinations}
\alias{icombinations}
\title{Combinations iterator}
\usage{
icombinations(n, k, x = NULL, f = NULL, replace = FALSE)
}
\arguments{
\item{n}{an integer, would be determined implicitly from `x` or `f` if missing}

\item{k}{an integer}

\item{x}{an optional vector indicating item labels}

\item{f}{an integer vector of item repeat frequencies}

\item{replace}{an logical to select with replacement}
}
\description{
This function returns a \href{Combinations-class.html}{Combinations} object which
allows users to fetch the combinations via the \code{getnext} method.
}
\seealso{
\link{combinations} for generating combinations and \link{ncombinations} to calculate number of combinations
}
