% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{ipartitions}
\alias{ipartitions}
\title{Partitions iterator}
\usage{
ipartitions(n, k = NULL, descending = FALSE)
}
\description{
This function returns a \href{Combinations-class.html}{Combinations} object which
allows users to fetch the partitions via the \code{getnext} method.
}
