% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{ipermutations}
\alias{ipermutations}
\title{Permutations generator}
\usage{
ipermutations(n, k = n, x = NULL, f = NULL, replace = FALSE)
}
\arguments{
\item{n}{an integer, would be determined implicitly from \code{x} or \code{f} if missing}

\item{k}{an integer}

\item{x}{an optional vector indicating item labels}

\item{f}{an integer vector of item repeat frequencies}

\item{replace}{an logical to select with replacement}
}
\description{
Permutations generator
}
