% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{Partitions-class}
\alias{Partitions-class}
\title{Partitions class}
\description{
An R6 class of partitions iterator. \link{ipartitions} is a convenient wrapper for initializing the class.
}
\section{Usage}{

\preformatted{
Partitions$new(n, k, x = NULL, f = NULL, replace = FALSE)
...$getnext(d = 1L, type = "r", drop = d == 1L)
...$collect(type = "r")
}
}

\seealso{
\link{ipartitions}
}
