% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{Partitions-class}
\alias{Partitions-class}
\title{Partitions class}
\arguments{
\item{n}{integer to be partitioned}

\item{m}{integer: number of partitions}

\item{descending}{logical: lexicographical or reverse lexicographical order}
}
\description{
\preformatted{
Partitions$new(n, m=NULL, descending = FALSE)
}
}
