% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\docType{data}
\name{Permutations-class}
\alias{Permutations-class}
\alias{Permutations}
\title{Permutations class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Permutations
}
\description{
An R6 class of permutation iterator. \link{ipermutations} is a convenient wrapper for initializing the class.
}
\section{Methods}{

\preformatted{
Permutations$new(n, k, x = NULL, f = NULL, replace = FALSE)
.$getnext(d = 1L, type = "r", drop = d == 1L)
.$collect(type = "r")
}
}

\seealso{
\link{ipermutations}
}
\keyword{datasets}
