% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{Permutations-class}
\alias{Permutations-class}
\title{Permutations class}
\description{
An R6 class of permutation iterator. \link{ipermutations} is a convenient wrapper for initializing the class.
}
\seealso{
\link{ipermutations}
}
